test: MockBoxOfficeInfo 수정
