+++ BoxofficeInfoTests/BoxofficeInfoTests.swift
@@ -10,11 +10,11 @@ import XCTest
 
 final class BoxofficeInfoTests: XCTestCase {
     
-    var sut: BoxofficeInfo<DailyBoxofficeObject>!
+    var sut: MockBoxofficeInfo<DailyBoxofficeObject>!
     var mockSession = MockURLSession()
 
     override func setUpWithError() throws {
-        sut = BoxofficeInfo(interfaceValue: "20230320", apiType: .boxoffice, session: mockSession)
+        sut = MockBoxofficeInfo(interfaceValue: "20230320", apiType: .boxoffice, session: mockSession)
     }
 
     override func tearDownWithError() throws {
@@ -24,7 +24,6 @@ final class BoxofficeInfoTests: XCTestCase {
     func test_boxofficeType값은_일별_박스오피스이다() {
         // given
         let expectation = "일별 박스오피스"
-    
         // when
         // then
         sut.search { event in
